backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.9.0
slowapi==0.1.9
openai>=1.0.0
